[refactor]: hover 추가 및 컴포넌트 분리 & 검색 결과 hooks 생성
