+++ src/app/search/_components/SearchBar.module.css
@@ -1,0 +1,38 @@
+.searchContainer {
+  margin-bottom: 24px;
+}
+
+.searchInputContainer {
+  display: flex;
+  gap: 10px;
+  width: 100%;
+  margin: 0 auto;
+}
+
+.searchInput {
+  flex: 1;
+  padding: 12px 16px;
+  border: 2px solid #e0e0e0;
+  border-radius: 8px;
+  font-size: 16px;
+  outline: none;
+  transition: border-color 0.2s;
+}
+
+.searchInput:focus {
+  border-color: #ff6f61;
+}
+
+.searchButton {
+  padding: 12px 16px;
+  background-color: #ff6f61;
+  color: white;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+.searchButton:hover {
+  background-color: #e55a4f;
+}
