+++ src/app/search/page.module.css
@@ -1,130 +1,5 @@
 .container {
-  height: 100%;
   width: 100%;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.searchContainer {
-  margin-bottom: 24px;
-}
-
-.searchInputContainer {
-  display: flex;
-  gap: 10px;
-  width: 100%;
-  margin: 0 auto;
-}
-
-.searchInput {
-  flex: 1;
-  padding: 12px 16px;
-  border: 2px solid #e0e0e0;
-  border-radius: 8px;
-  font-size: 16px;
-  outline: none;
-  transition: border-color 0.2s;
-}
-
-.searchInput:focus {
-  border-color: #ff6f61;
-}
-
-.searchButton {
-  padding: 12px 16px;
-  background-color: #ff6f61;
-  color: white;
-  border: none;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.searchButton:hover {
-  background-color: #e55a4f;
-}
-
-.searchResultContainer {
   display: flex;
   flex-direction: column;
-  gap: 36px;
-}
-
-.searchResultItem {
-  display: flex;
-  gap: 15px;
-  background-color: white;
-  align-items: center;
-  cursor: pointer;
-}
-
-.searchResultItemImage {
-  border-radius: 50%;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 64px;
-  height: 64px;
-}
-
-.searchResultItemContent h3 {
-  margin: 0 0 8px 0;
-  font-size: 20px;
-  color: #333;
-  font-weight: 600;
-}
-
-.searchResultItemContent p {
-  margin: 0 0 8px 0;
-  color: #666;
-  font-size: 14px;
-}
-
-.searchResultType {
-  display: inline-block;
-  padding: 4px 8px;
-  background-color: #ff6f61;
-  color: white;
-  border-radius: 4px;
-  font-size: 12px;
-  font-weight: 500;
-  width: fit-content;
-}
-
-/* 축제별 게시글 그룹 스타일 */
-.festivalPostGroup {
-  display: flex;
-  flex-direction: row;
-  gap: 16px;
-  background-color: white;
-  align-items: center;
-  cursor: pointer;
-}
-
-.festivalPostGroupCount {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 60px;
-  height: 60px;
-  border-radius: 50%;
-  background-color: #ffffff;
-  border: 4px solid #ff6f61;
-  color: #ff6f61;
-  font-size: 42px;
-  font-weight: 800;
-}
-
-.festivalPostGroupInfo {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 8px;
-}
-
-.festivalPostGroupInfo h3 {
-  font-size: 20px;
-  color: #333;
-  font-weight: 600;
 }
